:recycle: simplifying and abstracting the LocatorConfig initialization, in order to move the jaxb initialization out of GuiceLocator in the next commit
