+++ consumer/telemetry-app/src/main/resources/locator.xml
@@ -14,13 +14,6 @@
  -->
 <!DOCTYPE xml>
 <locator-config>
-    <provided>
-        <provide>
-            <interceptor>annotation</interceptor>
-            <with>Impl</with>
-        </provide>
-    </provided>
-
     <packages>
         <package>org.eclipse.kapua.consumer.telemetry</package>
         <package>org.eclipse.kapua.commons</package>
