+++ extras/encryption-migrator/src/main/resources/locator.xml
@@ -13,8 +13,6 @@
  -->
 <!DOCTYPE xml>
 <locator-config>
-    <provided/>
-
     <packages>
         <package>org.eclipse.kapua.commons</package>
         <package>org.eclipse.kapua.service</package>
