+++ job-engine/app/web/src/main/resources/locator.xml
@@ -13,9 +13,6 @@
  -->
 <!DOCTYPE xml>
 <locator-config>
-    <provided>
-    </provided>
-
     <packages>
         <package>org.eclipse.kapua.job.engine.app.web</package>
         <package>org.eclipse.kapua.plugin</package>
