+++ locator/guice/src/main/java/org/eclipse/kapua/locator/guice/KapuaModule.java
@@ -13,6 +13,27 @@
  *******************************************************************************/
 package org.eclipse.kapua.locator.guice;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.annotation.PostConstruct;
+
+import org.aopalliance.intercept.MethodInterceptor;
+import org.eclipse.kapua.KapuaException;
+import org.eclipse.kapua.commons.core.AbstractKapuaModule;
+import org.eclipse.kapua.commons.core.InterceptorBind;
+import org.eclipse.kapua.commons.core.ServiceModule;
+import org.eclipse.kapua.commons.core.ServiceModuleBundle;
+import org.eclipse.kapua.locator.KapuaProvider;
+import org.eclipse.kapua.locator.LocatorConfig;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.google.common.collect.ImmutableSet;
 import com.google.common.reflect.ClassPath;
 import com.google.common.reflect.ClassPath.ClassInfo;
@@ -25,65 +46,27 @@ import com.google.inject.multibindings.ProvidesIntoSet;
 import com.google.inject.spi.InjectionListener;
 import com.google.inject.spi.TypeEncounter;
 import com.google.inject.spi.TypeListener;
-import org.aopalliance.intercept.MethodInterceptor;
-import org.eclipse.kapua.KapuaErrorCodes;
-import org.eclipse.kapua.KapuaException;
-import org.eclipse.kapua.KapuaRuntimeException;
-import org.eclipse.kapua.commons.core.AbstractKapuaModule;
-import org.eclipse.kapua.commons.core.InterceptorBind;
-import org.eclipse.kapua.commons.core.ServiceModule;
-import org.eclipse.kapua.commons.core.ServiceModuleBundle;
-import org.eclipse.kapua.commons.util.ResourceUtils;
-import org.eclipse.kapua.locator.KapuaProvider;
-import org.eclipse.kapua.model.KapuaObjectFactory;
-import org.eclipse.kapua.service.KapuaService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.annotation.PostConstruct;
-import java.lang.annotation.Annotation;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.net.URL;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 
 public class KapuaModule extends AbstractKapuaModule {
 
     private static final Logger logger = LoggerFactory.getLogger(KapuaModule.class);
+    private final LocatorConfig locatorConfig;
 
-    /**
-     * Service resource file from which the managed services are read
-     */
-    private static final String SERVICE_RESOURCE = "locator.xml";
-
-    private final String resourceName;
-
-    public KapuaModule(final String resourceName) {
-        this.resourceName = resourceName;
+    public KapuaModule(final LocatorConfig locatorConfig) {
+        this.locatorConfig = locatorConfig;
     }
 
     @Override
     protected void configureModule() {
+        bind(LocatorConfig.class).toInstance(locatorConfig);
+
+        // Packages are supposed to contain service implementations
+        Collection<String> packageNames = locatorConfig.getIncludedPackageNames();
+
+        // Packages that are excluded
+        Collection<String> excludedPkgNames = locatorConfig.getExcludedPackageNames();
+
         try {
-            // Find locator configuration file
-            List<URL> locatorConfigurations = Arrays.asList(ResourceUtils.getResource(resourceName));
-            if (locatorConfigurations.isEmpty()) {
-                return;
-            }
-
-            // Read configurations from resource files
-            final LocatorConfig locatorConfig = new LocatorConfigurationExtractorImpl(locatorConfigurations.get(0)).fetchLocatorConfig();
-
-            // Packages are supposed to contain service implementations
-            Collection<String> packageNames = locatorConfig.getIncludedPackageNames();
-
-            // Packages that are excluded
-            Collection<String> excludedPkgNames = locatorConfig.getExcludedPackageNames();
-
             ClassLoader classLoader = this.getClass().getClassLoader();
             ClassPath classPath = ClassPath.from(classLoader);
             boolean initialize = true;
@@ -101,60 +84,12 @@ public class KapuaModule extends AbstractKapuaModule {
                     }
                     logger.trace("CLASS: {}", classInfo.getName());
                     Class<?> theClass = Class.forName(classInfo.getName(), !initialize, classLoader);
+
                     KapuaProvider serviceProvider = theClass.getAnnotation(KapuaProvider.class);
                     if (serviceProvider != null) {
                         providers.add(theClass);
                     }
                 }
-            }
-
-            // Provided names are the objects provided by the module (services or factories
-            Collection<String> providedInterfaceNames = locatorConfig.getProvidedInterfaceNames();
-
-            for (String providedName : providedInterfaceNames) {
-
-                boolean isClassBound = false;
-
-                final String trimmedServiceLine = providedName.trim();
-                Class<?> kapuaObject = Class.forName(trimmedServiceLine, !initialize, classLoader);
-
-                // When the provided object is a service ...
-                // ... add binding with a matching implementation
-                if (KapuaService.class.isAssignableFrom(kapuaObject)) {
-                    for (Class<?> clazz : providers) {
-                        if (kapuaObject.isAssignableFrom(clazz)) {
-                            ServiceResolver<KapuaService, ?> resolver = ServiceResolver.newInstance(kapuaObject, clazz);
-                            bind(resolver.getServiceClass()).to(resolver.getImplementationClass()).in(Singleton.class);
-                            logger.info("Bind Kapua service {} to {}", kapuaObject, clazz);
-                            isClassBound = true;
-                            break;
-                        }
-                    }
-
-                    if (isClassBound) {
-                        continue;
-                    }
-                }
-
-                // When the provided object is a factory ...
-                // ... add binding with a matching implementation
-                if (KapuaObjectFactory.class.isAssignableFrom(kapuaObject)) {
-                    for (Class<?> clazz : providers) {
-                        if (kapuaObject.isAssignableFrom(clazz)) {
-                            FactoryResolver<KapuaObjectFactory, ?> resolver = FactoryResolver.newInstance(kapuaObject, clazz);
-                            bind(resolver.getFactoryClass()).to(resolver.getImplementationClass()).in(Singleton.class);
-                            logger.info("Bind Kapua factory {} to {}", kapuaObject, clazz);
-                            isClassBound = true;
-                            break;
-                        }
-                    }
-
-                    if (isClassBound) {
-                        continue;
-                    }
-                }
-
-                logger.warn("No provider found for {}", kapuaObject);
             }
 
             // Bind interceptors
@@ -180,16 +115,15 @@ public class KapuaModule extends AbstractKapuaModule {
             bind(ServiceModuleBundle.class).in(Singleton.class);
 
             bindListener(new HasPostConstructAnnotationMatcher(), new TypeListener() {
+
                 @Override
                 public <I> void hear(TypeLiteral<I> type, TypeEncounter<I> encounter) {
                     encounter.register(PostConstructAnnotationInvoker.INSTANCE);
                 }
             });
             logger.trace("Binding completed");
-
         } catch (Exception e) {
-            logger.error("Exeption configuring module", e);
-            throw new KapuaRuntimeException(KapuaErrorCodes.INTERNAL_ERROR, e, "Cannot load " + SERVICE_RESOURCE);
+            throw new RuntimeException(e);
         }
     }
 
@@ -206,8 +140,8 @@ public class KapuaModule extends AbstractKapuaModule {
         }
     }
 
-
     private static class PostConstructAnnotationInvoker implements InjectionListener<Object> {
+
         private static final PostConstructAnnotationInvoker INSTANCE = new PostConstructAnnotationInvoker();
 
         private boolean hasPostConstructAnnotation(Method method) {
@@ -237,6 +171,7 @@ public class KapuaModule extends AbstractKapuaModule {
     @ProvidesIntoSet
     ServiceModule voidServiceModule() {
         return new ServiceModule() {
+
             @Override
             public void start() throws KapuaException {
