+++ locator/guice/src/main/java/org/eclipse/kapua/locator/guice/LocatorConfigurationExtractorImpl.java
@@ -13,11 +13,25 @@
  *******************************************************************************/
 package org.eclipse.kapua.locator.guice;
 
-import org.eclipse.kapua.commons.util.ResourceUtils;
-
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+
+import org.apache.commons.configuration.ConfigurationException;
+import org.apache.commons.configuration.XMLConfiguration;
+import org.eclipse.kapua.commons.util.ResourceUtils;
+import org.eclipse.kapua.locator.KapuaLocatorErrorCodes;
+import org.eclipse.kapua.locator.KapuaLocatorException;
+import org.eclipse.kapua.locator.LocatorConfig;
+import org.eclipse.kapua.locator.LocatorConfigurationExtractor;
 
 public class LocatorConfigurationExtractorImpl implements LocatorConfigurationExtractor {
+
+    private static final String SERVICE_RESOURCE_INCLUDED_PACKAGES = "packages.package";
+    private static final String SERVICE_RESOURCE_EXCLUDED_PACKAGES = "packages.excludes.package";
+
     private final LocatorConfig locatorConfig;
 
     public LocatorConfigurationExtractorImpl(String locatorConfigName) throws Exception {
@@ -28,7 +42,46 @@ public class LocatorConfigurationExtractorImpl implements LocatorConfigurationEx
         if (locatorConfigUrl == null) {
             throw new Exception("Locator configuration cannot be found: " + locatorConfigUrl);
         }
-        this.locatorConfig = LocatorConfig.fromURL(locatorConfigUrl);
+        List<String> includedPkgNames = new ArrayList<>();
+        List<String> excludedPkgNames = new ArrayList<>();
+
+        XMLConfiguration xmlConfig;
+        try {
+            xmlConfig = new XMLConfiguration(locatorConfigUrl);
+        } catch (ConfigurationException e) {
+            throw new KapuaLocatorException(KapuaLocatorErrorCodes.INVALID_CONFIGURATION, e);
+        }
+
+        Object props = xmlConfig.getProperty(SERVICE_RESOURCE_INCLUDED_PACKAGES);
+        if (props instanceof Collection<?>) {
+            addAllStrings(includedPkgNames, (Collection<?>) props);
+        }
+        if (props instanceof String) {
+            includedPkgNames.add((String) props);
+        }
+
+        props = xmlConfig.getProperty(SERVICE_RESOURCE_EXCLUDED_PACKAGES);
+        if (props instanceof Collection<?>) {
+            addAllStrings(excludedPkgNames, (Collection<?>) props);
+        }
+        if (props instanceof String) {
+            excludedPkgNames.add((String) props);
+        }
+
+        this.locatorConfig = new LocatorConfig(Collections.unmodifiableList(includedPkgNames), Collections.unmodifiableList(excludedPkgNames));
+    }
+
+    /**
+     * @param list
+     * @param other
+     * @since 1.0.0
+     */
+    private static void addAllStrings(List<String> list, Collection<?> other) {
+        for (Object entry : other) {
+            if (entry instanceof String) {
+                list.add((String) entry);
+            }
+        }
     }
 
     @Override
