+++ qa/integration/src/test/resources/locator.xml
@@ -13,12 +13,6 @@
  -->
 <!DOCTYPE xml>
 <locator-config>
-    <provided>
-        <provide>
-            <interceptor>annotation</interceptor>
-            <with>Impl</with>
-        </provide>
-    </provided>
     <packages>
         <package>org.eclipse.kapua.integration.misc</package>
         <package>org.eclipse.kapua.plugin</package>
