+++ service/authentication-app/src/main/resources/locator.xml
@@ -14,12 +14,6 @@
  -->
 <!DOCTYPE xml>
 <locator-config>
-    <provided>
-        <provide>
-            <interceptor>annotation</interceptor>
-            <with>Impl</with>
-        </provide>
-    </provided>
     <packages>
         <package>org.eclipse.kapua.service.authentication</package>
         <package>org.eclipse.kapua.commons</package>
